requirements.txt

fastapi==0.114.2
uvicorn==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.9
pytest==8.2.1
httpx==0.27.2
coverage==7.6.1
